requirements.txt
Flask==2.3.2
grpcio==1.56.0
grpcio-tools==1.56.0
gunicorn==20.1.0
numpy==1.25.1
opencv-python-headless==4.8.0.74
PyYAML==6.0
